reduced PageViews.kt lines and moved pages to its respective features
