+++ app/src/main/java/com/ilustris/cuccina/feature/recipe/domain/service/RecipeService.kt
@@ -62,4 +62,5 @@ class RecipeService : BaseService() {
     }
 
     suspend fun getRecipesByUser(userID: String) = query(userID, "userID")
+    suspend fun getRecipesByCategory(category: String) = query(category, "category")
 }
