+++ app/src/main/java/com/ilustris/cuccina/feature/recipe/step/presentation/ui/StepsPageView.kt
@@ -1,0 +1,39 @@
+package com.ilustris.cuccina.feature.recipe.step.presentation.ui
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import com.ilustris.cuccina.ui.theme.Page
+
+@Composable
+fun StepsPageView(page: Page.StepsPage) {
+    LazyColumn(modifier = Modifier.fillMaxSize(), verticalArrangement = Arrangement.Center) {
+        item {
+            Text(
+                text = page.title,
+                modifier = Modifier.fillMaxWidth(),
+                style = MaterialTheme.typography.headlineLarge.copy(textAlign = TextAlign.Center)
+            )
+            Text(
+                text = page.description,
+                modifier = Modifier.fillMaxWidth(),
+                style = MaterialTheme.typography.labelMedium.copy(textAlign = TextAlign.Center)
+            )
+        }
+
+        items(page.steps.size) { index ->
+            StepItem(
+                step = page.steps[index],
+                canEdit = false,
+                ingredients = emptyList(),
+                updateStep = {})
+        }
+
+    }
+}
