+++ app/src/main/java/com/ilustris/cuccina/ui/theme/Components.kt
@@ -1,28 +1,31 @@
 package com.ilustris.cuccina.ui.theme
 
+import ai.atick.material.MaterialColor
+import android.util.Log
 import androidx.compose.animation.core.*
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawWithCache
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.BlendMode
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.graphics.*
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.airbnb.lottie.compose.*
 import com.ilustris.cuccina.R
 import com.ilustris.cuccina.feature.recipe.ui.component.StateComponent
 import com.silent.ilustriscore.core.model.ViewModelBaseState
@@ -33,7 +36,9 @@ fun appColors() =
     listOf(
         MaterialTheme.colorScheme.primary,
         MaterialTheme.colorScheme.secondary,
-        MaterialTheme.colorScheme.tertiary
+        MaterialTheme.colorScheme.primary.copy(alpha = 0.5f),
+        MaterialTheme.colorScheme.tertiary,
+        MaterialTheme.colorScheme.secondary.copy(alpha = 0.5f),
     )
 
 @Composable
@@ -60,7 +65,7 @@ fun CuccinaLoader() {
         val brush = Brush.linearGradient(
             appColors(),
             start = Offset(offsetAnimation.value, offsetAnimation.value),
-            end = Offset(x = offsetAnimation.value * 3, y = offsetAnimation.value)
+            end = Offset(x = offsetAnimation.value * 5, y = offsetAnimation.value * 3)
         )
         Icon(
             imageVector = ImageVector.vectorResource(id = R.drawable.cherry),
@@ -86,7 +91,6 @@ fun CuccinaLoader() {
 
     }
 }
-
 
 @Composable
 fun getStateComponent(state: ViewModelBaseState, action: (ViewModelBaseState) -> Unit) {
@@ -126,5 +130,201 @@ fun getStateComponent(state: ViewModelBaseState, action: (ViewModelBaseState) ->
 @Composable
 fun CuccinaLoaderPreview() {
     CuccinaLoader()
+}
+
+
+fun annotatedPage(text: String, annotations: List<String>, color: Color) = buildAnnotatedString {
+    append(text)
+    Log.i("PageAnnotation", "annotatedPage: validating annotations -> $annotations on ($text)")
+    annotations.forEach { annotation ->
+        if (text.contains(annotation, true)) {
+            val startIndex = text.indexOf(annotation)
+            val endIndex = text.indexOf(annotation) + annotation.length
+            if (startIndex != -1 && endIndex != text.length) {
+                Log.i(javaClass.simpleName, "annotation: adding style to $annotation")
+                addStyle(
+                    SpanStyle(
+                        color = color,
+                        fontWeight = FontWeight.Bold
+                    ),
+                    start = text.indexOf(annotation),
+                    end = (text.indexOf(annotation) + annotation.length)
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun SimplePageView(page: Page.SimplePage) {
+
+    Column(
+        modifier = Modifier
+            .background(page.backColor ?: MaterialTheme.colorScheme.background)
+            .padding(16.dp)
+            .fillMaxSize(), verticalArrangement = Arrangement.Center
+    ) {
+        val textColor = page.textColor ?: MaterialTheme.colorScheme.onBackground
+        Text(
+            text = page.title,
+            modifier = Modifier.fillMaxWidth(),
+            style = MaterialTheme
+                .typography
+                .headlineLarge
+                .copy(textAlign = TextAlign.Center, color = textColor)
+        )
+        Text(
+            text = annotatedPage(
+                page.description,
+                page.annotatedTexts,
+                MaterialTheme.colorScheme.primary
+            ),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 8.dp),
+            style = MaterialTheme
+                .typography
+                .bodyMedium
+                .copy(textAlign = TextAlign.Center, color = textColor)
+        )
+    }
+}
+
+@Composable
+fun AnimatedTextPage(page: Page.AnimatedTextPage) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(page.backColor ?: MaterialTheme.colorScheme.background)
+    ) {
+        var texts = page.texts.toString().replace("[", "").replace("]", "").replace(",", "")
+        val textColor = page.textColor ?: MaterialTheme.colorScheme.onBackground
+
+        repeat(3 * page.texts.size) {
+            texts += texts
+        }
+
+        Text(
+            texts,
+            letterSpacing = 3.sp,
+            lineHeight = 50.sp,
+            style = MaterialTheme.typography.headlineLarge,
+            modifier = Modifier
+                .graphicsLayer {
+                    val scale = 2.5f
+                    scaleX = scale
+                    scaleY = scale
+                }
+                .fillMaxSize()
+        )
+
+
+
+        Column(
+            verticalArrangement = Arrangement.Center,
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp)
+        ) {
+
+            Text(
+                text = page.title,
+                modifier = Modifier.fillMaxWidth(),
+                style = MaterialTheme
+                    .typography
+                    .headlineLarge
+                    .copy(
+                        textAlign = TextAlign.Center,
+                        color = textColor,
+                        fontWeight = FontWeight.Bold,
+                        shadow = Shadow(
+                            color = MaterialColor.Black,
+                            offset = Offset(1f, 1f),
+                            blurRadius = 1.3f
+                        )
+                    )
+            )
+            Text(
+                text = page.description,
+                modifier = Modifier.fillMaxWidth(),
+                style = MaterialTheme
+                    .typography
+                    .bodyMedium
+                    .copy(
+                        textAlign = TextAlign.Center,
+                        shadow = Shadow(
+                            color = MaterialColor.Black,
+                            offset = Offset(1f, 1f),
+                            blurRadius = 1.3f
+                        ),
+                        color = textColor
+                    )
+            )
+
+        }
+
+
+    }
+}
+
+@Composable
+fun SuccessPageView(page: Page.SuccessPage, pageAction: () -> Unit) {
+    Column(
+        modifier = Modifier
+            .background(page.backColor ?: MaterialTheme.colorScheme.background)
+            .padding(16.dp)
+            .fillMaxSize(), verticalArrangement = Arrangement.Center
+    ) {
+        val textColor = page.textColor ?: MaterialTheme.colorScheme.onBackground
+
+        val celebrateComposition by rememberLottieComposition(
+            LottieCompositionSpec.RawRes(R.raw.happytoast)
+        )
+
+        val celebrateProgress by animateLottieCompositionAsState(
+            celebrateComposition,
+            isPlaying = true,
+            iterations = LottieConstants.IterateForever
+        )
+
+        LottieAnimation(
+            celebrateComposition,
+            celebrateProgress,
+            modifier = Modifier
+                .fillMaxWidth()
+                .fillMaxHeight(0.3f)
+        )
+
+
+        Text(
+            text = page.title,
+            modifier = Modifier.fillMaxWidth(),
+            style = MaterialTheme
+                .typography
+                .headlineLarge
+                .copy(textAlign = TextAlign.Center, color = textColor)
+        )
+        Text(
+            text = page.description,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 8.dp),
+            style = MaterialTheme
+                .typography
+                .bodyMedium
+                .copy(textAlign = TextAlign.Center, color = textColor)
+        )
+
+        Button(
+            onClick = pageAction,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 16.dp)
+        ) {
+            Text(
+                text = page.actionText
+            )
+        }
+    }
 }
